# src/services/report/routes/v1/public/doc/openapi.json
{
  "openapi": "3.0.0",
...
  "info": {
    "title": "ezReeport API",
    "version": "1.0",
    "contact": {
      "name": "ezTeam",
      "url": "https://github.com/ezpaarse-project",
      "email": "ezpaarse@couperin.org"
    },
    "license": {
      "name": "CeCILL",
      "url": "https://github.com/ezpaarse-project/ezreeport/blob/master/LICENSE.txt"
    },
    "description": "Reporting service"
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Direct"
    },
    {
      "url": "/report/api/v1",
      "description": "ezMESURE"
    }
  ],
  "paths": {
...
    "/templates": {
      "get": {
        "summary": "Get templates",
        "tags": [
          "templates"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": [
                        {
                          "name": "basic",
                          "renderer": "vega-pdf",
                          "pageCount": 6
                        },
                        {
                          "name": "tests/slot-tests",
                          "renderer": "vega-pdf",
                          "pageCount": 9
                        }
                      ]
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Template"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "pageCount": {
                                "type": "number"
                              }
                            }
                          }
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "default": {
                          "type": "string",
                          "description": "Id of the default template"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "templates-get",
        "security": [
          {
            "User Token": []
          }
        ],
        "description": "Get all templates usable by tasks.\n\nNeeds `general.templates-get` (see `/me/permissions` for more info)."
      },
      "parameters": [],
      "post": {
        "summary": "Create template",
        "operationId": "post-templates",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "apiVersion": 1,
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "scratch",
                        "tags": [],
                        "body": {
                          "layouts": []
                        },
                        "createdAt": "2023-04-11T09:31:24.620Z",
                        "updatedAt": "2023-04-11T09:31:24.620Z",
                        "renderer": "vega-pdf",
                        "pageCount": 0
                      }
                    }
                  },
                  "properties": {
                    "apiVersion": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/TemplateFull"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Add a template usable by tasks.\n\nNeeds `general.templates-post` (see `/me/permissions` for more info).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TemplateFull"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          },
          "description": ""
        },
        "security": [
          {
            "User Token": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
    "/templates/{templateId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "templateId",
          "in": "path",
          "required": true,
          "description": "The template id"
        }
      ],
      "get": {
        "summary": "Get specific template",
        "tags": [
          "templates"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "tests/slot-tests",
                        "renderer": "vega-pdf",
                        "pageCount": 9,
                        "template": {}
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/TemplateFull"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "operationId": "templates-get-name(*)",
        "security": [
          {
            "User Token": []
          }
        ],
        "parameters": [],
        "description": "Get specific template.\n\nNeeds `general.templates-get-template` (see `/me/permissions` for more info).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFull"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update or create specific template",
        "operationId": "put-templates-layoutName",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "scratch",
                        "tags": [],
                        "body": {
                          "layouts": []
                        },
                        "createdAt": "2023-04-11T09:31:24.620Z",
                        "updatedAt": "2023-04-11T09:31:24.620Z",
                        "renderer": "vega-pdf",
                        "pageCount": 0
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/TemplateFull"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-examples": {
                    "Example 1": {
                      "status": {
                        "code": 200,
                        "message": "OK"
                      },
                      "content": {
                        "name": "scratch",
                        "tags": [],
                        "body": {
                          "layouts": []
                        },
                        "createdAt": "2023-04-11T09:31:24.620Z",
                        "updatedAt": "2023-04-11T09:31:24.620Z",
                        "renderer": "vega-pdf",
                        "pageCount": 0
                      }
                    }
                  },
                  "properties": {
                    "status": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    },
                    "content": {
                      "$ref": "#/components/schemas/TemplateFull"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Update a template if found, create if not.\n\nNeeds `general.templates-put-template` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFull"
              }
            }
          }
        },
        "tags": [
          "templates"
        ]
      },
      "delete": {
        "summary": "Delete specific template",
        "operationId": "delete-templates-layoutName",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        },
        "description": "Delete template.\n\nNeeds `general.templates-delete-template` (see `/me/permissions` for more info).",
        "security": [
          {
            "User Token": []
          }
        ],
        "tags": [
          "templates"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "x-examples": {},
        "description": "Reporting task",
        "x-tags": [
          "task"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string"
          },
          "namespaceId": {
            "type": "string",
            "description": "Namespace concerned by the task",
            "readOnly": true
          },
          "recurrence": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "MONTHLY",
              "QUARTERLY",
              "BIENNIAL",
              "YEARLY"
            ],
            "description": "Defines when the task will be run."
          },
          "nextRun": {
            "type": "string",
            "description": "Next time the task will be run. Calculated if omiited",
            "format": "date-time"
          },
          "lastRun": {
            "type": "string",
            "format": "date-time",
            "description": "Last time the task has run",
            "nullable": true,
            "readOnly": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Is the generation of the task is enabled",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "name",
          "recurrence"
        ]
      },
...
      "Template": {
        "title": "Template",
        "x-stoplight": {
          "id": "gadmdr9m2pb87"
        },
        "type": "object",
        "additionalProperties": false,
        "description": "Report template with a few more info",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "color": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TemplateFull": {
        "title": "TemplateFull",
        "x-stoplight": {
          "id": "f249696c2d47f"
        },
        "type": "object",
        "additionalProperties": false,
        "description": "Full info about template",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "readOnly": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "color": {
                  "type": "string",
                  "nullable": true
                }
              }
            }
          },
          "body": {
            "$ref": "#/components/schemas/ReportTemplate"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "pageCount": {
            "type": "number",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "required": [
          "name",
          "tags",
          "body"
        ]
      },
      "ReportTemplate": {
        "title": "ReportTemplate",
        "x-stoplight": {
          "id": "ioqkkh3sl06p9"
        },
        "type": "object",
        "additionalProperties": false,
        "description": "Base template to use when generating tasks",
        "properties": {
          "layouts": {
            "$ref": "#/components/schemas/ReportLayout"
          },
          "fetchOptions": {
            "type": "object"
          },
          "renderer": {
            "type": "string",
            "description": "Renderer needed to render this template",
            "nullable": true
          },
          "renderOptions": {
            "type": "object",
            "description": "Options passed to the renderer"
          }
        }
      },
      "ReportLayout": {
        "title": "ReportLayout",
        "x-stoplight": {
          "id": "13fa278fffb4e"
        },
        "type": "object",
        "description": "Layout info. Can have more properties",
        "x-examples": {
          "Example": {
            "name": "basic",
            "renderer": "vega-pdf",
            "pageCount": 6
          }
        },
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "string"
              }
            ]
          },
          "fetcher": {
            "type": "string",
            "description": "Fetcher used to fetch data",
            "nullable": true
          },
          "fetchOptions": {
            "type": "object",
            "description": "Options passed to the fetcher"
          },
          "figures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportFigure"
            }
          }
        },
        "required": [
          "figures"
        ]
      },
      "ReportFigure": {
        "title": "ReportFigure",
        "x-stoplight": {
          "id": "b0c34fa35eb7c"
        },
        "type": "object",
        "description": "Reporting figure",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "Type of figure"
          },
          "data": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "object"
                    },
                    {
                      "type": "array",
                      "items": {}
                    },
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    }
                  ]
                }
              }
            ]
          },
          "slots": {
            "type": "array",
            "description": "Slot used by figure",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "type",
          "data"
        ]
      },
...
    },
    "securitySchemes": {
      "User Token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used by user to interact with service"
      },
      "API Key": {
        "name": "X-API-Key",
        "type": "apiKey",
        "in": "header",
        "description": "Used by linked application to manage service"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "JWT malformed: invalid signature"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 400,
                    "message": "Bad Request"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'GET /users/' requires auth"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 401,
                    "message": "Unauthorized"
                  },
                  "content": {
                    "message": "'...' requires auth"
                  }
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User 'user' doesn't have the rights to acces to 'GET /users/'"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 403,
                    "message": "Forbidden"
                  },
                  "content": {
                    "message": "User '...' doesn't have the rights to access to '...'"
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "User 'testa' not found"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 404,
                    "message": "Not Found"
                  },
                  "content": {
                    "message": "... '...' not found"
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "content": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "x-examples": {
                "example-1": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "Unexpected token v in JSON at position 0"
                  }
                }
              }
            },
            "examples": {
              "Example": {
                "value": {
                  "status": {
                    "code": 500,
                    "message": "Internal Server Error"
                  },
                  "content": {
                    "message": "..."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}